Split out binary work arounds into a dep1.cpp/h
