+++ source/compiler-core/slang-llvm-compiler.cpp
@@ -1,6 +1,8 @@
 // slang-llvm-compiler.cpp
 #include "slang-llvm-compiler.h"
 
+#include "slang-downstream-dep1.h"
+
 #include "../core/slang-common.h"
 #include "../../slang-com-helper.h"
 
@@ -21,24 +23,6 @@
 
 namespace Slang
 {
-
-class DownstreamCompilerAdapter_Dep1 : public DownstreamCompilerBase
-{
-public:
-    // IDownstreamCompiler
-    virtual SLANG_NO_THROW const Desc& SLANG_MCALL getDesc() SLANG_OVERRIDE { return m_dep->getDesc(); }
-    virtual SLANG_NO_THROW SlangResult SLANG_MCALL compile(const CompileOptions& options, RefPtr<DownstreamCompileResult>& outResult) SLANG_OVERRIDE { return m_dep->compile(options, outResult); }
-    virtual SLANG_NO_THROW SlangResult SLANG_MCALL disassemble(SlangCompileTarget sourceBlobTarget, const void* blob, size_t blobSize, ISlangBlob** out) SLANG_OVERRIDE { return m_dep->disassemble(sourceBlobTarget, blob, blobSize, out); }
-    virtual SLANG_NO_THROW bool SLANG_MCALL isFileBased() SLANG_OVERRIDE { return m_dep->isFileBased(); }
-
-    DownstreamCompilerAdapter_Dep1(DownstreamCompiler_Dep1* dep): 
-        m_dep(dep)
-    {
-    }
-
-protected:
-    RefPtr<DownstreamCompiler_Dep1> m_dep;
-};
 
 /* static */SlangResult LLVMDownstreamCompilerUtil::locateCompilers(const String& path, ISlangSharedLibraryLoader* loader, DownstreamCompilerSet* set)
 {
